Fix font selector broken on Pharo 9 and Pharo 10
On Pharo9 FreeTypeFontSelectorDialogWindow is still used to select PTerm font.
On Pharo10 however, this class no longer exists, FontChooser is used instead.

FontChooser exists on Pharo9, but the class interface is completely different
from Pharo10. It is impossible to have a compatible use of this class for both
Pharo9 and Pharo10